initial docfile, "no_index" support
First shot at docfile finding code...  Very basic, needs work.

Additionally, we now look at META.yml for a "no_index" section, and
generate the corresponding filter macro entries.  The entire macros generation
section bugs me, and really needs to be refactored so we track what needs to
be filtered in this class, and the actual implementation of the filtering can
be done in the spec tt2.