Initial fully-managed spec support
Here we have a combination of two efforts:  one, the fully-managed-prep
command, that takes a spec and splits the various sections out into an
auto.ini file; and two, the related updates to SpecData and descendents
required to use these.  ...one essential step closer to having the spec
file be fully managed by maintainer tools.